Refactor parts of Parser into Executor
