+++ src/main/java/duke/main/Parser.java
@@ -3,44 +3,16 @@ package duke.main;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
-import duke.task.TaskList;
 import duke.task.ToDo;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 /**
  * Handles the parsing of user commands.
  */
 public class Parser {
-    private Storage storage;
-    private Ui ui;
-    private TaskList taskList;
-
-    /**
-     * Constructor for Parser.
-     *
-     * @param storage  Storage from Duke.
-     * @param ui       Ui from Duke.
-     * @param taskList TaskList from Duke.
-     */
-    public Parser(Storage storage, Ui ui, TaskList taskList) {
-        this.storage = storage;
-        this.ui = ui;
-        this.taskList = taskList;
-    }
-
-    /**
-     * Parses user inputs.
-     *
-     * @param input String from user.
-     * @return String response from the parser.
-     */
-    public String parseAndExecute(String input) {
-        String command = getCommandFromInput(input);
-        String description = getDescriptionFromInput(input);
-        String response = executeCommand(command, description);
-        return response;
-    }
 
     /**
      * Returns the command in user input.
@@ -48,7 +20,7 @@ public class Parser {
      * @param input String from user.
      * @return String command.
      */
-    private String getCommandFromInput(String input) {
+    public static String getCommandFromInput(String input) {
         String[] commandAndDesc = splitInput(input);
         return commandAndDesc[0];
     }
@@ -59,7 +31,7 @@ public class Parser {
      * @param input String from user.
      * @return String description.
      */
-    private String getDescriptionFromInput(String input) {
+    public static String getDescriptionFromInput(String input) {
         String[] commandAndDesc = splitInput(input);
         boolean containsDescription = commandAndDesc.length == 2;
         return containsDescription ? commandAndDesc[1] : "";
@@ -71,100 +43,10 @@ public class Parser {
      * @param input String from user.
      * @return String[] containing parts of the input.
      */
-    private String[] splitInput(String input) {
+    public static String[] splitInput(String input) {
         return input.split(" ", 2);
     }
 
-    /**
-     * Executes the command provided.
-     *
-     * @param command     String type of command.
-     * @param description String description of the command.
-     * @return String response upon executing the command.
-     */
-    private String executeCommand(String command, String description) {
-        String response;
-
-        //@formatter:off
-        switch (command) {
-        case "bye":
-            response = ui.exitWithGoodbye();
-            break;
-        case "find":
-            List<Task> matches = taskList.findTasks(description);
-            response = ui.getMatchingTasksSummary(matches);
-            break;
-        case "tag":
-            String[] taskNumAndTag = description.split(" ", 2);
-            String taskNum = taskNumAndTag[0];
-
-            if (taskNumAndTag.length < 2) {
-                throw new DukeException("\tPlease enter a tag");
-            }
-
-            String tag = taskNumAndTag[1];
-            Task toTag = selectTask(taskNum);
-            response = toTag.addTag(tag);
-            break;
-        case "clear":
-            String resetTaskMessage = storage.resetTasks();
-            String clearTaskMessage = taskList.clearTasks();
-            response = resetTaskMessage + clearTaskMessage;
-            break;
-        case "list":
-            response = ui.getTaskListSummary(taskList);
-            break;
-        case "deadline":
-            response = taskList.addTask(new Deadline(description));
-            break;
-        case "event":
-            response = taskList.addTask(new Event(description));
-            break;
-        case "todo":
-            response = taskList.addTask(new ToDo(description));
-            break;
-        case "done":
-            Task selectedTask = selectTask(description);
-            response = selectedTask.markAsDone();
-            break;
-        case "delete":
-            Task toDelete = selectTask(description);
-            response = taskList.deleteTask(toDelete);
-            break;
-        default:
-            if (command.equals("")) {
-                response = ui.getEmptyInputMessage();
-            } else {
-                response = ui.getUnknownCommandMessage(command);
-            }
-        }
-        //@formatter:on
-
-        updateStorage();
-        return response;
-    }
-
-    /**
-     * Commits the changes in taskList to storage.
-     */
-    private void updateStorage() {
-        storage.writeToStorage(taskList);
-    }
-
-    /**
-     * Selects the task with the given description.
-     *
-     * @param description String number corresponding to the task.
-     * @return Task selected.
-     */
-    private Task selectTask(String description) {
-        int taskNum = getTaskNum(description);
-        Task selectedTask = taskList.getTask(taskNum);
-        if (selectedTask == null) {
-            throw new DukeException("\tSorry, I can't seem to find that task\n");
-        }
-        return selectedTask;
-    }
 
     /**
      * Parses the description for the task number.
@@ -172,11 +54,65 @@ public class Parser {
      * @param description String containing the task number.
      * @return int task number.
      */
-    private int getTaskNum(String description) {
+    public static int getTaskNum(String description) {
         try {
             return Integer.parseInt(description);
         } catch (NumberFormatException e) {
             throw new DukeException("\tI'm Sorry, WHAT?!?!\n");
         }
+    }
+
+    /**
+     * Parses the task string and returns the corresponding task.
+     *
+     * @param line String for a task retrieved from storage.
+     * @return Task that is parsed.
+     */
+    public static Task generateTaskFromLine(String line) {
+        //task storage string formatted as taskType | isComplete | description | time (optional)
+        String[] fragments = line.split(" \\| ");
+
+        String taskType = fragments[0];
+        boolean isComplete = Boolean.parseBoolean(fragments[1]);
+        String description = fragments[2];
+        String tags = fragments[3];
+        List<String> tagList = new ArrayList<>(Arrays.asList(tags.split("#")));
+        String time = fragments[4];
+
+        return generateTaskFromDetails(taskType, isComplete, description, time, tagList);
+    }
+
+    /**
+     * Generate a task from the given parameters.
+     *
+     * @param taskType        String type of task.
+     * @param isCompletedTask boolean true if task is completed, else false.
+     * @param taskDescription String description of the task.
+     * @param time            String time associated with the task.
+     * @param tags            list of Tags for the task.
+     * @return Task generated.
+     */
+    private static Task generateTaskFromDetails(String taskType, boolean isCompletedTask, String taskDescription,
+                                                String time,
+                                                List<String> tags) {
+        Task foundTask;
+
+        //@formatter:off
+        switch (taskType) {
+        case "T":
+            foundTask = new ToDo(taskDescription, isCompletedTask, tags);
+            break;
+        case "D":
+            foundTask = new Deadline(taskDescription, time, isCompletedTask, tags);
+            break;
+        case "E":
+            foundTask = new Event(taskDescription, time, isCompletedTask, tags);
+            break;
+        default:
+            throw new DukeException("\t OOPS!!! I can't find your tasks.\n");
+        }
+        //@formatter:on
+
+        return foundTask;
     }
 }
