+++ src/main/java/duke/main/Storage.java
@@ -1,17 +1,12 @@
 package duke.main;
 
-import duke.task.Deadline;
-import duke.task.Event;
 import duke.task.Task;
 import duke.task.TaskList;
-import duke.task.ToDo;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -39,7 +34,7 @@ public class Storage {
     public TaskList loadTaskList() {
         try {
             Stream<String> taskString = Files.lines(filePath);
-            List<Task> taskList = taskString.map(this::loadTaskFromString).collect(Collectors.toList());
+            List<Task> taskList = taskString.map(Parser::generateTaskFromLine).collect(Collectors.toList());
             return new TaskList(taskList);
         } catch (IOException e) {
             throw new DukeException("\t OOPS!!! I can't find your tasks.\n");
@@ -47,59 +42,6 @@ public class Storage {
             resetTasks();
             throw new DukeException("\t OOPS!!! Your tasks might be corrupted.");
         }
-    }
-
-    /**
-     * Parses the task string and returns the corresponding task.
-     *
-     * @param line String for a task retrieved from storage.
-     * @return Task that is parsed.
-     */
-    private Task loadTaskFromString(String line) {
-        //task storage string formatted as taskType | isComplete | description | time (optional)
-        String[] fragments = line.split(" \\| ");
-
-        String taskType = fragments[0];
-        boolean isComplete = Boolean.parseBoolean(fragments[1]);
-        String description = fragments[2];
-        String tags = fragments[3];
-        List<String> tagList = new ArrayList<>(Arrays.asList(tags.split("#")));
-        String time = fragments[4];
-
-        return generateTask(taskType, isComplete, description, time, tagList);
-    }
-
-    /**
-     * Generate a task from the given parameters.
-     *
-     * @param taskType        String type of task.
-     * @param isCompletedTask boolean true if task is completed, else false.
-     * @param taskDescription String description of the task.
-     * @param time            String time associated with the task.
-     * @param tags            list of Tags for the task.
-     * @return Task generated.
-     */
-    private Task generateTask(String taskType, boolean isCompletedTask, String taskDescription, String time,
-                              List<String> tags) {
-        Task foundTask;
-
-        //@formatter:off
-        switch (taskType) {
-        case "T":
-            foundTask = new ToDo(taskDescription, isCompletedTask, tags);
-            break;
-        case "D":
-            foundTask = new Deadline(taskDescription, time, isCompletedTask, tags);
-            break;
-        case "E":
-            foundTask = new Event(taskDescription, time, isCompletedTask, tags);
-            break;
-        default:
-            throw new DukeException("\t OOPS!!! I can't find your tasks.\n");
-        }
-        //@formatter:on
-
-        return foundTask;
     }
 
     /**
